Use OpenNetworkConnection instead of calling ConnectNode directly in Dynamic code
ConnectNode is meant to be private and should not be leaked outside of
CConnman. This leaking resulted in multiple problems with AddRef/Release
behaviour. We use OpenNetworkConnection/OpenDynodeConnection from now on
and use appropriate ForNode calls to access the CNode* instance.

This commit also changes the behaviour for already connected nodes which
are afterwards requested for a dynode connection. We don't set
fDynode=true anymore on such nodes, which means it won't be auto
disconnected anymore from ProcessDynodeConnections.